`Scopes.all` returns a `List` object instead of an array of items
Some breaking changes:

**1. `Cubscout::Scopes.all` returns a `Cubscout::List`**

By returning an array of items, we were losing the ability to query the metadata:
page number, total (non-paginated) number of elements found by the query, total
number of pages. We can now query those as part of the returned `Cubscout::List`
object.

By returning a `Cubscout::List` object, we can still iterate over it's items as
the class includes `Enumerable` and was already meant to work this way.

This promotes `Cubscout::List` as a public API class now. So it will be
harder to change it's API.

**2. Cubscout::List method name changes**

`Cubscout::List#size` becomes `Cubscout::List#total_size`: `size` should reflect
the number of elements that can be iterated over. In this case it should
correspond to `Cubscout::List#page_size`, but experience shows it does not
always match. Because of that, `Cubscout::List#size` is a separate method and
always returns the number of `Cubscout::List#items`.
`Cubscout::List#total_size` is the total amount (non-paginated) of items that
correspond to the query.

This is a breaking change, but `Cubscout::List` was private API so far.